minor refactor to break out getObstaclesFromCircuitJson
